[infrastructure] Migrate NTP to bnd
