+++ CODEOWNERS
@@ -88,8 +88,6 @@
 /addons/binding/org.openhab.binding.netatmo/ @clinique @cweitkamp @lolodomo
 /addons/binding/org.openhab.binding.nibeuplink/ @alexf2015
 /addons/binding/org.openhab.binding.nikohomecontrol/ @mherwege
-/addons/binding/org.openhab.binding.ntp/ @marcelrv
-/addons/binding/org.openhab.binding.ntp.test/ @marcelrv
 /addons/binding/org.openhab.binding.nuki/ @mkatter
 /addons/binding/org.openhab.binding.oceanic/ @kgoderis
 /addons/binding/org.openhab.binding.onebusaway/ @sdwilsh
@@ -204,6 +202,7 @@
 /bundles/org.openhab.binding.nest/ @wborn
 /bundles/org.openhab.binding.network/ @davidgraeff @mettke
 /bundles/org.openhab.binding.nibeheatpump/ @paulianttila
+/bundles/org.openhab.binding.ntp/ @marcelrv
 /bundles/org.openhab.binding.onewire/ @J-N-K
 /bundles/org.openhab.binding.openuv/ @clinique
 /bundles/org.openhab.binding.openweathermap/ @cweitkamp
