--- addons/binding/org.openhab.binding.ntp.test/META-INF/MANIFEST.MF
@@ -1,29 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: NTP Binding Tests
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-SymbolicName: org.openhab.binding.ntp.test
-Bundle-Vendor: openHAB
-Bundle-Version: 2.5.0.qualifier
-Fragment-Host: org.openhab.binding.ntp
-Import-Package: 
- org.apache.commons.io,
- org.apache.commons.lang,
- org.eclipse.jdt.annotation;resolution:=optional,
- org.eclipse.smarthome.config.discovery,
- org.eclipse.smarthome.core.common.registry,
- org.eclipse.smarthome.core.events,
- org.eclipse.smarthome.core.items,
- org.eclipse.smarthome.core.items.events,
- org.eclipse.smarthome.core.library.items,
- org.eclipse.smarthome.core.thing.link,
- org.eclipse.smarthome.test,
- org.eclipse.smarthome.test.java,
- org.eclipse.smarthome.test.storage,
- org.hamcrest;core=split,
- org.junit,
- org.mockito,
- org.mockito.stubbing,
- org.osgi.framework,
- org.osgi.service.component
-Automatic-Module-Name: org.openhab.binding.ntp.test
