--- addons/binding/org.openhab.binding.ntp.test/build.properties
@@ -1,5 +1,0 @@
-source.. = src/test/java/
-output.. = target/test-classes/
-bin.includes = META-INF/,\
-               .,\
-               NOTICE
