--- addons/binding/org.openhab.binding.ntp/ESH-INF/binding/binding.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<binding:binding id="ntp" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:binding="http://eclipse.org/smarthome/schemas/binding/v1.0.0"
-	xsi:schemaLocation="http://eclipse.org/smarthome/schemas/binding/v1.0.0 http://eclipse.org/smarthome/schemas/binding-1.0.0.xsd">
-
-	<name>NTP Binding</name>
-	<description>The NTP Binding polls the configured timeserver and post the current date and time.</description>
-	<author>Marcel Verpaalen</author>
-
-</binding:binding>
