--- addons/binding/org.openhab.binding.ntp/META-INF/MANIFEST.MF
@@ -1,25 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ClassPath: .
-Bundle-ManifestVersion: 2
-Bundle-Name: NTP Binding
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-SymbolicName: org.openhab.binding.ntp;singleton:=true
-Bundle-Vendor: openHAB
-Bundle-Version: 2.5.0.qualifier
-Import-Package: 
- org.apache.commons.net,
- org.apache.commons.net.ntp,
- org.eclipse.jdt.annotation;resolution:=optional,
- org.eclipse.smarthome.config.core,
- org.eclipse.smarthome.config.discovery,
- org.eclipse.smarthome.core.i18n,
- org.eclipse.smarthome.core.items,
- org.eclipse.smarthome.core.library.types,
- org.eclipse.smarthome.core.thing,
- org.eclipse.smarthome.core.thing.binding,
- org.eclipse.smarthome.core.thing.binding.builder,
- org.eclipse.smarthome.core.thing.type,
- org.eclipse.smarthome.core.types,
- org.slf4j
-Service-Component: OSGI-INF/*.xml
-Automatic-Module-Name: org.openhab.binding.ntp
