--- addons/binding/org.openhab.binding.ntp/OSGI-INF/.gitignore
@@ -1,1 +1,0 @@
-*.xml
