--- addons/binding/org.openhab.binding.ntp/build.properties
@@ -1,7 +1,0 @@
-source.. = src/main/java/
-output.. = target/classes
-bin.includes = META-INF/,\
-               .,\
-               OSGI-INF/,\
-               ESH-INF/,\
-               NOTICE
