--- addons/binding/org.openhab.binding.ntp/pom.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.openhab.binding</groupId>
-    <artifactId>pom</artifactId>
-    <version>2.5.0-SNAPSHOT</version>
-  </parent>
-
-  <groupId>org.openhab.binding</groupId>
-  <artifactId>org.openhab.binding.ntp</artifactId>
-  <packaging>eclipse-plugin</packaging>
-
-  <name>NTP Binding</name>
-
-</project>
