+++ bundles/org.openhab.binding.ntp/NOTICE
