+++ bundles/org.openhab.binding.ntp/src/main/resources/ESH-INF/i18n/ntp_de.properties
