+++ features/karaf/openhab-addons/src/main/feature/feature.xml
@@ -523,7 +523,7 @@
 
     <feature name="openhab-binding-ntp" description="NTP Binding" version="${project.version}">
         <feature>openhab-runtime-base</feature>
-        <bundle start-level="80">mvn:org.openhab.binding/org.openhab.binding.ntp/${project.version}</bundle>
+        <bundle start-level="80">mvn:org.openhab.addons.bundles/org.openhab.binding.ntp/${project.version}</bundle>
     </feature>
 
     <feature name="openhab-binding-nuki" description="Nuki Binding" version="${project.version}">
